Add min width to header top
Set the min width to the same value as the elements below it so that
it can have the exact same amount of space for styling. Add gap to
header links and header center for improved spacing. Set header top to
space between for ideal spacing in that flex container.