Add styling to buttons
Add padding to create space between the letters inside of the button
and the edges of the button. Add margin so the button doesn't touch
other elements. Set border to none and add border radius to round the
corners of the button. Use pixels specifically because percentage was
not rounding the border properly.