Remove H2 and quote and more from container
Remove H2, quote, and call to action from container. Needed a div to
select all of the sub containers to be able to style them all at once
in the same flex container. Body div is also selecting everything from
H2 to footer, so having container do almost the same is unecessary.